requirements.txt
alembic==0.9.6
Babel==2.5.1
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-CKEditor==0.3.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
httplib2==0.10.3
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.10
MarkupSafe==1.0
oauth2==1.9.0.post1
passlib==1.7.1
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
pytz==2017.3
regex==2017.12.12
ruamel.yaml==0.15.35
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.0
tzlocal==1.5.1
Unidecode==0.4.21
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
